# server/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MapApp",
        "version": "1.0",
        "description": "An API for application, which helps delivery companies to plan their routes.",
        "x-logo": {
            "url": ""
        }
    },
    "servers": [
        {
            "url": "http://localhost:8081/api/v2",
            "description": "Local server"
        }
    ],
    "paths": {
        "/data/client": {
            "summary": "Client",
            "description": "Client is an object representing the client of the delivery company. It has a type field, which may have only \"recipient\" or \"sender\" values. Type \"sender\" is a company from which the parcels are being shipped. Type \"recipient\" is a person/another company to whom a parcels are delivered.\nNotice that only registered users having created a Profile can save their Clients' data.",
            "get": {
                "tags": [
                    "client"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/200"
                                },
                                "examples": {
                                    "client": {
                                        "value": {
                                            "data": [
                                                {
                                                    "id": 1,
                                                    "username": "user1",
                                                    "name": "User",
                                                    "type": "sender",
                                                    "city": "Helsinki",
                                                    "street": "Ruusulankatu",
                                                    "building": "4B",
                                                    "flat": 12,
                                                    "lat": 24.9245931,
                                                    "lon": 60.1815267
                                                },
                                                {
                                                    "id": 2,
                                                    "username": "user2",
                                                    "name": "User",
                                                    "type": "recipient",
                                                    "city": "Lahti",
                                                    "street": "Leantie",
                                                    "building": "19",
                                                    "flat": 67,
                                                    "lat": 25.6664544,
                                                    "lon": 60.9702099
                                                }
                                            ],
                                            "metadata": {
                                                "currentPage": 3,
                                                "limit": 2
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Successfully requested"
                    },
                    "401": {
                        "$ref": "#/components/responses/401",
                        "x-last-modified": 1719224545166
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500",
                        "x-last-modified": 1719224573070
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Read all Clients",
                "description": "Read all Client associated with the logged-in user Profile. It is possible to request only own Clients\n\n| Requirement / Function | Is Required / Has |\n| -------------------------- | ------------------  |\n| Authentication                | yes |\n| Authorization                  | only own  |\n| Pagination                       | yes |\n| Search                             | yes |"
            },
            "put": {
                "requestBody": {
                    "description": "Notice that only the id field and  fields, that need to be updated should be provided",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "client"
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204",
                        "x-last-modified": 1718890060914
                    },
                    "400": {
                        "$ref": "#/components/responses/400",
                        "x-last-modified": 1718890078434
                    },
                    "401": {
                        "$ref": "#/components/responses/401",
                        "x-last-modified": 1718890117079
                    },
                    "404": {
                        "$ref": "#/components/responses/404",
                        "x-last-modified": 1718890145614
                    },
                    "409": {
                        "$ref": "#/components/responses/409",
                        "x-last-modified": 1718890158175
                    },
                    "500": {
                        "$ref": "#/components/responses/500",
                        "x-last-modified": 1718890182150
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update Client",
                "description": "| Requirement / Function | Is Required / Has |\n| ---------------------- | ----------------- |\n| Authentication         | yes |\n| Authorization          | only own |"
            },
            "post": {
                "requestBody": {
                    "description": "Notice that the Client will be automatically associated with the logged-in user profile",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientCreate"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "client"
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/201"
                                },
                                "examples": {
                                    "client": {
                                        "value": {
                                            "data": {
                                                "id": 1,
                                                "username": "user1",
                                                "name": "User",
                                                "type": "sender",
                                                "city": "Helsinki",
                                                "street": "Ruusulankatu",
                                                "building": "4B",
                                                "flat": 12,
                                                "lat": 24.9245931,
                                                "lon": 60.1815267
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Successfully created"
                    },
                    "400": {
                        "$ref": "#/components/responses/400",
                        "x-last-modified": 1718887948467
                    },
                    "401": {
                        "$ref": "#/components/responses/401",
                        "x-last-modified": 1718895012807
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "409": {
                        "$ref": "#/components/responses/409",
                        "x-last-modified": 1718887965493
                    },
                    "500": {
                        "$ref": "#/components/responses/500",
                        "x-last-modified": 1718887982927
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "operationId": "client_create",
                "summary": "Create client",
                "description": "| Requirement / Function | Is Required / Has |\n| -------------------------- | ------------------  |\n| Authentication                | yes |\n| Authorization                  | only for own Profile  |\n"
            }
        },
...
        "/data/client/{id}": {
            "summary": "Client",
            "description": "Client is an object representing the client of the delivery company. It has a type field, which may have only \"recipient\" or \"sender\" values. Type \"sender\" is a company from which the parcels are being shipped. Type \"recipient\" is a person/another company to whom a parcels are delivered.\nNotice that only registered users having created a Profile can save their Clients' data.",
            "get": {
                "tags": [
                    "client"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/200_one"
                                },
                                "examples": {
                                    "client": {
                                        "value": {
                                            "data": {
                                                "id": 1,
                                                "username": "user1",
                                                "name": "User",
                                                "type": "sender",
                                                "city": "Helsinki",
                                                "street": "Ruusulankatu",
                                                "building": "4B",
                                                "flat": 12,
                                                "lat": 24.9245931,
                                                "lon": 60.1815267
                                            }
                                        }
                                    }
                                }
                            }
                        },
                        "description": "Successfully requested"
                    },
                    "400": {
                        "$ref": "#/components/responses/400",
                        "x-last-modified": 1718889211678
                    },
                    "401": {
                        "$ref": "#/components/responses/401",
                        "x-last-modified": 1718889224633
                    },
                    "404": {
                        "$ref": "#/components/responses/404",
                        "x-last-modified": 1718889898774
                    },
                    "500": {
                        "$ref": "#/components/responses/500",
                        "x-last-modified": 1718889259586
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Read Client by id",
                "description": "| Requirement / Function | Is Required / Has |\n| -------------------------- | ------------------  |\n| Authentication                | yes |\n| Authorization                  | only own |\n| Pagination                       | no |\n| Search                             | no |"
            },
            "delete": {
                "tags": [
                    "client"
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204",
                        "x-last-modified": 1718889326604
                    },
                    "400": {
                        "$ref": "#/components/responses/400",
                        "x-last-modified": 1718889339851
                    },
                    "401": {
                        "$ref": "#/components/responses/401",
                        "x-last-modified": 1718889350835
                    },
                    "404": {
                        "$ref": "#/components/responses/404",
                        "x-last-modified": 1718889927760
                    },
                    "500": {
                        "$ref": "#/components/responses/500",
                        "x-last-modified": 1718889941430
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete Client by id",
                "description": "| Requirement / Function | Is Required / Has |\n| -------------------------- | ------------------  |\n| Authentication                | yes |\n| Authorization                  | only own |"
            },
            "parameters": [
                {
                    "examples": {
                        "id": {
                            "value": "74"
                        }
                    },
                    "name": "id",
                    "description": "client id",
                    "schema": {
                        "type": "integer"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
            "200": {
                "description": "\"data\"-field contains array of requested objects. \"metadata\"-field contains the pagination data",
                "required": [
                    "data",
                    "metadata"
                ],
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "metadata": {
                        "type": "object",
                        "properties": {
                            "currentPage": {
                                "description": "The current page number",
                                "type": "integer"
                            },
                            "limit": {
                                "description": "The limit of items per page",
                                "type": "integer"
                            }
                        }
                    }
                }
            },
            "201": {
                "description": "\"data\"-field contains created object.",
                "required": [
                    "data"
                ],
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object"
                    }
                }
            },
            "APIErrorRes": {
                "description": "Error response. Notice that errors field is always an array. Data can be ever empty object, null or undefined.",
                "type": "object",
                "properties": {
                    "data": {
                        "description": "An empty object for the data field",
                        "type": "object"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/APIError"
                        }
                    }
                },
                "x-last-modified": 1718628043172
            },
...
            "ClientCreate": {
                "description": "Client creation request body",
                "required": [
                    "username",
                    "type",
                    "city",
                    "street",
                    "building"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "description": "unique",
                        "type": "string",
                        "example": "client1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Client Name"
                    },
                    "type": {
                        "description": "recipient or sender",
                        "type": "string",
                        "example": "recipient"
                    },
                    "city": {
                        "type": "string",
                        "example": "Helsinki"
                    },
                    "street": {
                        "description": "street name only, without building",
                        "type": "string",
                        "example": "Punanotkonkatu"
                    },
                    "building": {
                        "description": "can be number or number with letter",
                        "type": "string",
                        "example": "3"
                    },
                    "flat": {
                        "description": "optional and not validated",
                        "type": "integer",
                        "example": 1
                    }
                }
            },
...
            "ClientUpdate": {
                "description": "Client update request body",
                "required": [
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "description": "read-only",
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "description": "unique",
                        "type": "string",
                        "example": "client1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Client Name"
                    },
                    "type": {
                        "description": "recipient or sender",
                        "type": "string",
                        "example": "recipient"
                    },
                    "city": {
                        "type": "string",
                        "example": "Helsinki"
                    },
                    "street": {
                        "description": "street name only, without building",
                        "type": "string",
                        "example": "Punanotkonkatu"
                    },
                    "building": {
                        "description": "can be number or number with letter",
                        "type": "string",
                        "example": "3"
                    },
                    "flat": {
                        "description": "optional and not validated",
                        "type": "integer",
                        "example": 1
                    }
                }
            },
...
            "200_one": {
                "description": "\"data\"-field contains requested object.",
                "required": [
                    "data"
                ],
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object"
                    }
                }
            }
        },
        "responses": {
...
            "204": {
                "description": "Successful request, response with no body. ",
                "x-last-modified": 1718887121189
            },
            "400": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIErrorRes"
                        },
                        "example": {
                            "errors": [
                                {
                                    "reason": "NOT_STRING",
                                    "status": 400,
                                    "location": "body",
                                    "message": "\"username\" field must be a string",
                                    "endpoint": "/api/v2/data/profile",
                                    "field": "Authorization",
                                    "name": "VALIDATION",
                                    "type": "APIError"
                                }
                            ]
                        }
                    }
                },
                "description": "Validation error response. Some of the fields are not specified, have wrong data type or otherwise in wrong format"
            },
            "401": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIErrorRes"
                        },
                        "example": {
                            "errors": [
                                {
                                    "reason": "AUTHENTICATION_FAILED",
                                    "status": 401,
                                    "location": "header",
                                    "message": "Could not authenticate with provided token. Token is invalid or expired",
                                    "endpoint": "/api/v2/data/profile",
                                    "field": "Authorization",
                                    "name": "NOT_AUTHENTICATED",
                                    "type": "APIError"
                                }
                            ]
                        }
                    }
                },
                "description": "Authentication error response. It may occur if the request does not contain \"Authorization\" header with the Bearer token",
                "x-last-modified": 1718629884243
            },
...
            "404": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIErrorRes"
                        },
                        "example": {
                            "errors": [
                                {
                                    "reason": "NOT_FOUND",
                                    "status": 404,
                                    "location": "param",
                                    "message": "Could not find any objects",
                                    "endpoint": "/api/v2/data/profile",
                                    "field": "id",
                                    "name": "NOT_FOUND",
                                    "type": "APIError"
                                }
                            ]
                        }
                    }
                },
                "description": "Nothing found",
                "x-last-modified": 1718889834399
            },
            "409": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIErrorRes"
                        },
                        "example": {
                            "errors": [
                                {
                                    "reason": "NOT_UNIQUE",
                                    "status": 409,
                                    "location": "body",
                                    "message": "User with that username already exists",
                                    "endpoint": "/api/v2/data/profile",
                                    "field": "username",
                                    "name": "BAD_REQUEST",
                                    "type": "APIError"
                                }
                            ]
                        }
                    }
                },
                "description": "Conflict error response. It may occur if some of the request field(s) is not unique",
                "x-last-modified": 1718631266638
            },
            "500": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIErrorRes"
                        },
                        "example": {
                            "errors": [
                                {
                                    "reason": "UNEXPECTED",
                                    "status": 500,
                                    "message": "Unexpected server error occurred",
                                    "endpoint": "/api/v2/data/profile",
                                    "additional": {
                                        "message": "throwAPIError is not defined",
                                        "stack": "ReferenceError: throwAPIError is not defined\n    at createProfile (file:///app/server/router/api/v2/data/profile.js:25:9)\n    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)\n    at async file:///app/server/router/api/v2/routeBuilder/core/pipelineHandlers/addController.js:12:30",
                                        "name": "ReferenceError"
                                    },
                                    "name": "UNEXPECTED",
                                    "type": "APIError"
                                }
                            ]
                        }
                    }
                },
                "description": "Server error response. It occurs unexpectedly. The additional field contains an error object",
                "x-last-modified": 1718631059302
            },
...
        },
...
        "securitySchemes": {
            "BearerAuth": {
                "scheme": "bearer",
                "type": "http",
                "description": "Bearer token"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
...
}